refactor(llm): separate translation context state
